+++ pdf/CHANGELOG.md
@@ -7,7 +7,7 @@
 - Add dashed lines to Decoration Widgets
 - Add TableRow decoration
 - Add Chart Widget [Marco Papula]
-- Add Divider Widget
+- Add Divider and VerticalDivider Widget
 - Replace Theme with ThemeData
 
 ## 1.6.2